map components structure & navbar visual
